+++ src/main/scala/com/rayrobdod/json/builder/SeqBuilder.scala
@@ -28,8 +28,8 @@ package com.rayrobdod.json.builder;
 
 import scala.collection.immutable.Seq
 import com.rayrobdod.json.parser.Parser
+import com.rayrobdod.json.union.ParserRetVal
 import com.rayrobdod.json.union.ParserRetVal.{Complex, Failure}
-import com.rayrobdod.json.union.NonPrimitiveParserRetVal
 
 /** 
  * A Builder that will build a Vector of values, where each inner value is produced by the parameter builder.
@@ -38,7 +38,7 @@ import com.rayrobdod.json.union.NonPrimitiveParserRetVal
  * 
  * [[#apply]] will return a left if the value is a primitive value.
  * 
- * @version 3.0
+ * @version 4.0
  * @tparam Key the type of keys encountered
  * @tparam Value the type of primitive values encountered
  * @tparam Inner the type of complex values produced by the childBuilder
@@ -48,7 +48,7 @@ import com.rayrobdod.json.union.NonPrimitiveParserRetVal
  */
 final class SeqBuilder[-Key, -Value, Inner](childBuilder:Builder[Key, Value, Inner]) extends Builder[Key, Value, Seq[Inner]] {
 	override def init:Seq[Inner] = Vector.empty[Inner]
-	override def apply[Input](folding:Seq[Inner], key:Key, innerInput:Input, parser:Parser[Key, Value, Input]):NonPrimitiveParserRetVal[Seq[Inner]] = {
+	override def apply[Input](folding:Seq[Inner], key:Key, innerInput:Input, parser:Parser[Key, Value, Input]):ParserRetVal[Seq[Inner], Nothing] = {
 		parser.parse(childBuilder, innerInput)
 				.complex.map{x => folding :+ x}
 				.primitive.flatMap{x => Failure("Primitive value in Seqbuilder", 0)}
@@ -64,11 +64,12 @@ final class SeqBuilder[-Key, -Value, Inner](childBuilder:Builder[Key, Value, Inn
  * [[#apply]] will return a left if the value is a complex value. 
  * 
  * @since 3.0
+ * @version 4.0
  * @tparam Value the type of primitive values encountered
  */
 final class PrimitiveSeqBuilder[Value] extends Builder[Any, Value, Seq[Value]] {
 	override def init:Seq[Value] = Vector.empty[Value]
-	override def apply[Input](folding:Seq[Value], key:Any, innerInput:Input, parser:Parser[Any, Value, Input]):NonPrimitiveParserRetVal[Seq[Value]] = {
+	override def apply[Input](folding:Seq[Value], key:Any, innerInput:Input, parser:Parser[Any, Value, Input]):ParserRetVal[Seq[Value], Nothing] = {
 		parser.parsePrimitive(innerInput).primitive.flatMap{x => Complex(folding :+ x)}.mergeToComplex
 	}
 }