+++ src/main/scala/com/rayrobdod/json/union/ParserRetVal.scala
@@ -26,7 +26,6 @@
 */
 package com.rayrobdod.json.union
 
-import scala.util.{Either, Left, Right}
 import com.rayrobdod.json.union.ParserRetVal.{ComplexProjection, PrimitiveProjection, FailureProjection}
 
 /*
@@ -40,7 +39,7 @@ import com.rayrobdod.json.union.ParserRetVal.{ComplexProjection, PrimitiveProjec
  * A union type representing possible return values of [[com.rayrobdod.json.parser.Parser.parse `Parser.parse`]].
  * 
  * @since 3.0
- * @version next
+ * @version 4.0
  */
 sealed trait ParserRetVal[+Complex, +Primitive] {
 	/**
@@ -57,19 +56,13 @@ sealed trait ParserRetVal[+Complex, +Primitive] {
 	/** projects this ParserRetVal as a complex */
 	def complex:ComplexProjection[Complex,Primitive] = new ComplexProjection(this)
 	/** projects this ParserRetVal as a failure
-	 * @since next
+	 * @since 4.0
 	 */
 	def failure:FailureProjection[Complex,Primitive] = new FailureProjection(this)
 	
-	/** Convert a Failure into a Left and the other two cases into a Right  */
-	def mergeToEither[A](implicit ev1:Primitive <:< A, ev2:Complex <:< A):Either[(String,Int),A] = {
-		this.fold({c => Right(ev2(c))}, {p => Right(ev1(p))}, {(s,i) => Left(s,i)})
-	}
-	def mergeToComplex[A](implicit ev1:Primitive <:< A, ev2:Complex <:< A):NonPrimitiveParserRetVal[A] = {
+	def mergeToComplex[A](implicit ev1:Primitive <:< A, ev2:Complex <:< A):ParserRetVal[A,Nothing] = {
 		this.fold({c => ParserRetVal.Complex(ev2(c))}, {p => ParserRetVal.Complex(ev1(p))}, {(s,i) => ParserRetVal.Failure(s,i)})
-	}
-	
-	/**
+	}	/**
 	 * @since 4.0
 	 */
 	def flip:ParserRetVal[Primitive, Complex] = this match {
@@ -122,18 +115,9 @@ sealed trait ParserRetVal[+Complex, +Primitive] {
 /**
  * A container for the types of [[ParserRetVal]]s
  * @since 3.0
- * @version 3.0.1
+ * @version 4.0
  */
 object ParserRetVal {
-	
-	/** Convert a Left to a Failure and a Right to a Complex */
-	def eitherToComplex[A](a:Either[(String,Int),A]):ParserRetVal[A,Nothing] = {
-		a.fold({case (msg,idx) => ParserRetVal.Failure(msg,idx)}, {x => ParserRetVal.Complex(x)})
-	}
-	/** Convert a Left to a Failure and a Right to a Primitive */
-	def eitherToPrimitive[A](a:Either[(String,Int),A]):ParserRetVal[Nothing,A] = {
-		a.fold({case (msg,idx) => ParserRetVal.Failure(msg,idx)}, {x => ParserRetVal.Primitive(x)})
-	}
 	
 	/** Represents a value that was produced by the parser without builder consultation */
 	final case class Primitive[+P](x:P) extends ParserRetVal[Nothing, P]{
@@ -141,7 +125,7 @@ object ParserRetVal {
 	}
 	
 	/** Represents a value that was produced via consultation of a builder */
-	final case class Complex[+C](x:C) extends ParserRetVal[C, Nothing] with NonPrimitiveParserRetVal[C] {
+	final case class Complex[+C](x:C) extends ParserRetVal[C, Nothing] {
 		def fold[Out](c:Function1[C,Out], p:Function1[Nothing,Out], f:Function2[String,Int,Out]):Out = c(x)
 	}
 	
@@ -152,7 +136,7 @@ object ParserRetVal {
 	 * @param idx the location in the input of the error. The meaning of idx depends on the Parser's Input;
 	 * 	if the Input is a character sequence, then idx might be the index of the character that caused a problem.
 	 */
-	final case class Failure(msg:String, idx:Int) extends ParserRetVal[Nothing, Nothing] with NonPrimitiveParserRetVal[Nothing] {
+	final case class Failure(msg:String, idx:Int) extends ParserRetVal[Nothing, Nothing] {
 		def fold[Out](c:Function1[Nothing,Out], p:Function1[Nothing,Out], f:Function2[String,Int,Out]):Out = f(msg,idx)
 	}
 	
@@ -172,13 +156,6 @@ object ParserRetVal {
 			case Primitive(p) => fun(p)
 			case c:Complex[C] => c
 			case f:Failure => f
-		}
-		
-		/** Return a Right if the backing value is a Primitive, else return a left */
-		def toEither:Either[(String,Int),P] = backing match {
-			case Primitive(p) => Right(p)
-			case c:Complex[C] => Left(("Expected primitive value", 0))
-			case Failure(msg, idx) => Left((msg, idx))
 		}
 	}
 	
@@ -200,13 +177,6 @@ object ParserRetVal {
 			case p:Primitive[P] => p
 			case f:Failure => f
 		}
-		
-		/** Return a Right if the backing value is a Complex, else return a left */
-		def toEither:Either[(String,Int),C] = backing match {
-			case Complex(c) => Right(c)
-			case p:Primitive[P] => Left(("Expected complex value", 0))
-			case Failure(msg, idx) => Left((msg, idx))
-		}
 	}
 	
 	/** A projection as if the ParserRetVal were a Failure
@@ -225,58 +195,4 @@ object ParserRetVal {
 		}
 	}
 	
-}
-
-/** 
- * @since next
- */
-sealed trait NonPrimitiveParserRetVal[+Success] extends ParserRetVal[Success, Nothing] {
-	import ParserRetVal.{Complex, Failure}
-	
-	def fold[Out](fc:Function1[Success,Out], ff:Function2[String,Int,Out]):Out = this match {
-		case Complex(c) => fc(c)
-		case Failure(m,i) => ff(m,i)
-	}
-	
-	/** projects this ParserRetVal as a complex */
-	override def complex:NonPrimitiveParserRetVal.ComplexProjection[Success] = new NonPrimitiveParserRetVal.ComplexProjection(this)
-	/** projects this ParserRetVal as a failure */
-	override def failure:NonPrimitiveParserRetVal.FailureProjection[Success] = new NonPrimitiveParserRetVal.FailureProjection(this)
-}
-
-/** 
- * @since next
- */
-object NonPrimitiveParserRetVal {
-	import ParserRetVal.{Complex, Failure}
-	
-	/** A projection as if the ParserRetVal were a Complex */
-	final class ComplexProjection[+C](backing:NonPrimitiveParserRetVal[C]) extends ParserRetVal.ComplexProjection(backing) {
-		
-		/** Map the backing value if the backing value is a Complex, else return the backing value */
-		override def map[X](fun:C => X):NonPrimitiveParserRetVal[X] = backing match {
-			case Complex(c) => Complex(fun(c))
-			case f:Failure => f
-		}
-		
-		/** Flatmap the backing value if the backing value is a Complex, else return the backing value
-		 */
-		def flatMap[X](fun:C => NonPrimitiveParserRetVal[X]):NonPrimitiveParserRetVal[X] = backing match {
-			case Complex(c) => fun(c)
-			case f:Failure => f
-		}
-	}
-	
-	/** A projection as if the ParserRetVal were a Complex */
-	final class FailureProjection[+C](backing:NonPrimitiveParserRetVal[C]) extends ParserRetVal.FailureProjection(backing) {
-		
-		/** Map the backing value if the backing value is a Failure, else return the backing value */
-		override def map(fun:(String, Int) => (String, Int)):NonPrimitiveParserRetVal[C] = backing match {
-			case c:Complex[C] => c
-			case Failure(msg, idx) => {
-				val (newMsg, newIdx) = fun.apply(msg, idx)
-				Failure(newMsg, newIdx)
-			}
-		}
-	}
 }